fix: Alarm 컴포넌트 디렉토리 변경 및 props 수정
